multilple roles for a user and file upload for user + admin view and display of all pending account

+++ WebApplication_Assignment SkillsLab2023/Models/TrainingModel.cs
@@ -9,13 +9,13 @@ namespace WebApplication_Assignment_SkillsLab2023.Models
     [Table("Training")]
     public class TrainingModel
     {
-        public int TrainingId {  get; set; }
+        public byte TrainingId {  get; set; }
         public string TrainingName { get; set; }
         public string TrainingDescription { get; set;}
         public string TrainingStatus { get; set; }
         public string DepartmentPriority {  get; set; }
-        public int SeatsAvailable {  get; set; }
-        public int SeatsTotal { get; set; }
+        public byte SeatsAvailable {  get; set; }
+        public byte SeatsTotal { get; set; }
         public DateTime TrainingRegistrationDeadline { get; set; }
     }
 }
